requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==5.0.7
django-compat==1.0.15
django-cors-headers==4.4.0
django-cron==0.6.0
django-extensions==3.2.3
django-oauth-toolkit==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
future==1.0.0
idna==3.7
jwcrypto==1.5.6
kombu==5.3.7
MarkupSafe==2.1.5
oauthlib==3.2.2
pillow==10.4.0
prompt_toolkit==3.0.47
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
serializers==0.2.4
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
